Move formula evaluator to separate file
This patch is done in order to decrease the size of the
`_formula_engine.py` file.
